With non-integer cycles, need more permutations
Previously used only permutations where the first timepoint was fixed,
since all timepoints are equivalent when you have an integer number of
cycles. But now if you have, say, 7 timepoints and period of 6, then the
first timepoint gets collapsed with the last, but all other timepoints
are 'alone'. So we cannot assume that the first timepoint is fixed.