Fix order of JTK/etc on comparisons
Since JTK and other stats are pulled from the computation backgend,
they are unaware of the comparisons and so must have the appropriate
rows selected from them. I.e. if the original spreadsheet contains 100
rows, a comparison with another spreadsheet may contain only some of
those rows and in a different order, but there are still 100 returned
JTK values and still in the original order. We force the ordering to
conform with the comaprison index instead